+++ src/main/kotlin/com/youquiz/authentication/global/config/SecurityConfiguration.kt
@@ -1,9 +1,12 @@
-package com.youquiz.template.global.config
+package com.youquiz.authentication.global.config
 
+import com.github.jwt.authentication.JwtAuthenticationFilter
+import com.github.jwt.core.JwtProvider
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.http.HttpStatus
 import org.springframework.security.config.annotation.web.reactive.EnableWebFluxSecurity
+import org.springframework.security.config.web.server.SecurityWebFiltersOrder
 import org.springframework.security.config.web.server.ServerHttpSecurity
 import org.springframework.security.web.server.SecurityWebFilterChain
 import org.springframework.security.web.server.authentication.HttpStatusServerEntryPoint
@@ -13,17 +16,16 @@ import org.springframework.security.web.server.context.NoOpServerSecurityContext
 @EnableWebFluxSecurity
 class SecurityConfiguration {
     @Bean
-    fun filterChain(http: ServerHttpSecurity): SecurityWebFilterChain =
+    fun filterChain(http: ServerHttpSecurity, jwtProvider: JwtProvider): SecurityWebFilterChain =
         with(http) {
             csrf { it.disable() }
-            formLogin { it.disable() }
-            logout { it.disable() }
             httpBasic { it.authenticationEntryPoint(HttpStatusServerEntryPoint(HttpStatus.UNAUTHORIZED)) }
             securityContextRepository(NoOpServerSecurityContextRepository.getInstance())
             authorizeExchange {
                 it.anyExchange()
                     .permitAll()
             }
+            addFilterAt(JwtAuthenticationFilter(jwtProvider), SecurityWebFiltersOrder.AUTHORIZATION)
             build()
         }
 }
